wsd/requirements.txt
django==5.1.3
pre-commit==2.9.2
iubeo==0.3.0
psycopg==3.2.3
django-lifecycle==1.0.0
django-extensions==3.1.0
django-hosts==5.1
django-admin-autocomplete-filter==0.7.1
python-dotenv==0.15.0
django-filter==2.4.0
django-pghistory==3.5.0
whitenoise==6.2.0
ume==0.1.1
pytz==2022.7
uritemplate==4.1.1
pyyaml==6.0
django-object-actions==4.1.0
django-allauth==65.2.0
django-sslserver==0.22
